+++ setup.py
@@ -38,12 +38,12 @@ setup(name='trimesh',
           'Topic :: Scientific/Engineering'],
       packages=[
           'trimesh',
-          'trimesh.io',
           'trimesh.ray',
           'trimesh.path',
-          'trimesh.path.io',
+          'trimesh.path.exchange',
           'trimesh.scene',
           'trimesh.visual',
+          'trimesh.exchange',
           'trimesh.resources',
           'trimesh.interfaces'],
       package_data={'trimesh': ['resources/*.template',
