+++ trimesh/path/exchange/dxf.py
