+++ trimesh/path/exchange/load.py
