+++ trimesh/path/exchange/misc.py
