+++ trimesh/path/exchange/svg_io.py
